form playground server actions

+++ src/actions.ts
@@ -5,6 +5,10 @@ import { prisma } from "./lib/prisma-connect";
 import { getServerSession } from "./lib/getServerSession";
 import { revalidatePath } from "next/cache";
 
+/**
+ * The function allows a user to create a new form
+ */
+
 export const startANewForm = async ({
   formId,
   title,
@@ -20,7 +24,7 @@ export const startANewForm = async ({
     await prisma.form.create({
       data: {
         id: formId,
-        title,
+        title: title.trim(),
         ownerId: session.user.id as string,
       },
     });
@@ -32,6 +36,10 @@ export const startANewForm = async ({
 
   revalidatePath("/forms");
 };
+
+/**
+ * The function allows a user to select one of the existing themes for a new form
+ */
 
 export const createNewFormTheme = async ({
   formId,
@@ -61,11 +69,30 @@ export const createNewFormTheme = async ({
   }
 };
 
-export const openRecentForm = async (formId: string) => {
+/**
+ * The function updates the last time a user opens a form
+ */
+
+export const openRecentForm = async ({
+  formId,
+  ownerId,
+}: {
+  formId: string;
+  ownerId: string;
+}) => {
   const session = await getServerSession();
 
   try {
     if (!session) return;
+    if (ownerId !== session.user.id) return;
+
+    const existingForm = await prisma.form.findUnique({
+      where: {
+        id: formId,
+      },
+    });
+
+    if (!existingForm) return;
 
     await prisma.form.update({
       where: {
@@ -85,6 +112,10 @@ export const openRecentForm = async (formId: string) => {
   redirect(`/forms/nx/${formId}/edit`);
 };
 
+/**
+ * The function allows a user to update form title
+ */
+
 export const updateFormTitle = async ({
   formId,
   title,
@@ -101,13 +132,21 @@ export const updateFormTitle = async ({
     if (ownerId !== session.user.id) return;
     if (title.replace(/\s+/g, "").length < 1) return;
 
+    const existingForm = await prisma.form.findUnique({
+      where: {
+        id: formId,
+      },
+    });
+
+    if (!existingForm) return;
+
     await prisma.form.update({
       where: {
         id: formId,
       },
 
       data: {
-        title: title,
+        title: title.trim(),
       },
     });
   } catch (error) {
@@ -119,6 +158,56 @@ export const updateFormTitle = async ({
   revalidatePath("/forms");
   revalidatePath(`/forms/nx/${formId}/edit`);
 };
+
+/**
+ * The function allows a user to update form description
+ */
+
+export const updateFormDescription = async ({
+  formId,
+  description,
+  ownerId,
+}: {
+  formId: string;
+  description: string;
+  ownerId: string;
+}) => {
+  const session = await getServerSession();
+
+  try {
+    if (!session) return;
+    if (ownerId !== session.user.id) return;
+    if (description.replace(/\s+/g, "").length < 1) return;
+
+    const existingForm = await prisma.form.findUnique({
+      where: {
+        id: formId,
+      },
+    });
+
+    if (!existingForm) return;
+
+    await prisma.form.update({
+      where: {
+        id: formId,
+      },
+
+      data: {
+        description: description.trim(),
+      },
+    });
+  } catch (error) {
+    console.error("error updating form title", error);
+  } finally {
+    await prisma.$disconnect();
+  }
+
+  revalidatePath(`/forms/nx/${formId}/edit`);
+};
+
+/**
+ * The function allows a user to delete a form
+ */
 
 export const deleteForm = async ({
   formId,
@@ -132,6 +221,14 @@ export const deleteForm = async ({
   try {
     if (!session) return;
     if (ownerId !== session.user.id) return;
+
+    const existingForm = await prisma.form.findUnique({
+      where: {
+        id: formId,
+      },
+    });
+
+    if (!existingForm) return;
 
     await prisma.theme.delete({
       where: {
@@ -151,4 +248,203 @@ export const deleteForm = async ({
   }
 
   revalidatePath("/forms");
+};
+
+// model Question {
+//   id                 String         @id @default(uuid())
+//   title              String?
+//   type               QUESTION_TYPE?
+//   required           Boolean        @default(false)
+//   multiChoiceOptions String[]
+//   checkboxesOptions  String[]
+//   dropDownOptions    String[]
+//   date               DateTime?
+//   time               DateTime?
+//   form               Form           @relation(fields: [formId], references: [id])
+//   formId             String
+// }
+
+/**
+ * The function allows a user to create/start a new question in a form
+ */
+
+export const createNewQuestion = async ({
+  questionId,
+  formId,
+  ownerId,
+}: {
+  questionId: string;
+  formId: string;
+  ownerId: string;
+}) => {
+  const session = await getServerSession();
+
+  try {
+    if (!session) return;
+    if (ownerId !== session.user.id) return;
+
+    await prisma.question.create({
+      data: {
+        id: questionId,
+        formId,
+      },
+    });
+  } catch (error) {
+    console.error("error deleting form", error);
+  } finally {
+    await prisma.$disconnect();
+  }
+};
+
+/**
+ * The function allows a user to create a short answer text question in a form
+ */
+
+export const createShortAnswerTextQuestion = async ({
+  formId,
+  title,
+  required,
+  questionId,
+  ownerId,
+}: {
+  formId: string;
+  title: string;
+  required: boolean;
+  questionId: string | undefined;
+  ownerId: string;
+}) => {
+  const session = await getServerSession();
+
+  try {
+    if (!session) return;
+    if (ownerId !== session.user.id) return;
+    if (title.replace(/\s+/g, "").length < 1) return;
+
+    const existingQuestion = await prisma.question.findUnique({
+      where: {
+        id: questionId,
+      },
+    });
+
+    if (!existingQuestion) return;
+
+    await prisma.question.update({
+      where: {
+        id: existingQuestion.id,
+      },
+      data: {
+        type: "SHORT_ANSWER",
+        formId,
+        title: title.trim(),
+        required,
+      },
+    });
+  } catch (error) {
+    console.error("error creating short answer question", error);
+  } finally {
+    await prisma.$disconnect();
+  }
+};
+
+/**
+ * The function allows a user to create a paragraph answer question in a form
+ */
+
+export const createParagraphAnswerQuestion = async ({
+  formId,
+  title,
+  required,
+  questionId,
+  ownerId,
+}: {
+  formId: string;
+  title: string;
+  required: boolean;
+  questionId: string | undefined;
+  ownerId: string;
+}) => {
+  const session = await getServerSession();
+
+  try {
+    if (!session) return;
+    if (ownerId !== session.user.id) return;
+    if (title.replace(/\s+/g, "").length < 1) return;
+
+    const existingQuestion = await prisma.question.findUnique({
+      where: {
+        id: questionId,
+      },
+    });
+
+    if (!existingQuestion) return;
+
+    await prisma.question.update({
+      where: {
+        id: existingQuestion.id,
+      },
+      data: {
+        type: "PARAGRAPH",
+        formId,
+        title: title.trim(),
+        required,
+      },
+    });
+  } catch (error) {
+    console.error("error creating paragraph answer question", error);
+  } finally {
+    await prisma.$disconnect();
+  }
+};
+
+/**
+ * The function allows a user to create a multi choice answer question in a form
+ */
+
+export const createMultiChoiceAnswerQuestion = async ({
+  formId,
+  title,
+  required,
+  multiChoiceOptions,
+  questionId,
+  ownerId,
+}: {
+  formId: string;
+  title: string;
+  required: boolean;
+  multiChoiceOptions: string[];
+  questionId: string | undefined;
+  ownerId: string;
+}) => {
+  const session = await getServerSession();
+
+  try {
+    if (!session) return;
+    if (ownerId !== session.user.id) return;
+    if (title.replace(/\s+/g, "").length < 1) return;
+
+    const existingQuestion = await prisma.question.findUnique({
+      where: {
+        id: questionId,
+      },
+    });
+
+    if (!existingQuestion) return;
+
+    await prisma.question.update({
+      where: {
+        id: existingQuestion.id,
+      },
+      data: {
+        type: "MULTIPLE_CHOICE",
+        formId,
+        title: title.trim(),
+        required,
+        multiChoiceOptions,
+      },
+    });
+  } catch (error) {
+    console.error("error deleting form", error);
+  } finally {
+    await prisma.$disconnect();
+  }
 };
